+++ codex-rs/external-agent-migration/src/detect/sessions/cla.rs
@@ -2,6 +2,7 @@ use super::common::SessionFileCandidate;
 use super::common::detect_recent_sessions;
 use crate::model::ExternalAgentSessionImportLimits;
 use crate::sessions::ExternalAgentSessionMigration;
+use crate::sessions::SessionRecordFormat;
 use std::fs;
 use std::io;
 use std::path::Path;
@@ -50,6 +51,7 @@ pub(crate) fn detect_recent_cla_sessions_with_limits(
             candidates.push(SessionFileCandidate {
                 path,
                 fallback_cwd: None,
+                record_format: SessionRecordFormat::Cla,
             });
         }
     }
